+++ example/components/TooltipExample.tsx
@@ -1,5 +1,5 @@
 import { LatLngTuple } from 'leaflet'
-import React, { useCallback, useMemo, useState } from 'react'
+import React, { useMemo, useState } from 'react'
 
 import {
   Circle,